requirement.txt
numpy
streamlit
python_dotenv
google_generativeai
pillow